% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyreg2.R
\name{getstv}
\alias{getstv}
\title{getstv}
\usage{
getstv(mod)
}
\arguments{
\item{mod}{modeloutput from hyreg2(...k = 1, latent = "both", ...)}
}
\value{
named vector
}
\description{
function to export coefficent values and names from a model fitted with k = 1 and latent = "both".
These values can be used as stv for a new model with k > 1
}
\examples{
formula <- y ~  -1 + x1 + x2 + x3 | id

k <- 1
stv <- setNames(c(0.2,0,1,1,1),c(colnames(simulated_data_norm)[3:5],c("sigma","theta")))
control = list(iter.max = 1000, verbose = 4)
rm(counter)
mod <- hyreg2(formula = formula,
                    data =  simulated_data_norm,
                    type =  simulated_data_norm$type,
                    stv = stv,
                    k = k,
                    type_cont = "TTO",
                    type_dich = "DCE_A",
                    opt_method = "L-BFGS-B",
                    control = control,
                    latent = "both",
                    id_col = "id"
)
new_stv <- getstv(mod)

### use new_stv in new model ###

#'mod <- hyreg2(formula = formula,
                    data =  simulated_data_norm,
                    type =  simulated_data_norm$type,
                    stv = new_stv,
                    k = 2,
                    type_cont = "TTO",
                    type_dich = "DCE_A",
                    opt_method = "L-BFGS-B",
                    control = control,
                    latent = "both",
                    id_col = "id"
}
\author{
Kim Rand & Svenja Elkenkamp
}
