% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hyreg2.R
\name{give_id}
\alias{give_id}
\title{give_id: function to decode which id or observation was classified to which class by the model}
\usage{
give_id(data, model, id = NULL)
}
\arguments{
\item{data}{a dataframe, which was used to estimate the model}

\item{model}{a flexmix model object estimtaed using hyreg2 or hyreg2_het}

\item{id}{one character, indicator for the column holding the ids, this column must be part of the provided dataframe.
the parameter must be specified, if the provided model was estimated under control for | id}
}
\value{
dataframe of two columns, first column named as provided id column or "observation" if id was not given as
         an input. second column named mod_comp indicating the assigned class for this id or observation
}
\description{
This function can be used to decode the classified classes by the model and see, which id ot observation
             was signed to which class
}
\details{
PUT DETAILS HERE
}
\author{
Svenja Elkenkamp & John Grosser
}
