% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hyreg2.R
\name{plot_hyreg2}
\alias{plot_hyreg2}
\title{plot_hyreg2: plot function to visualize the classification based on the model}
\usage{
plot_hyreg2(
  data,
  x,
  y,
  id_col,
  id_df_model,
  type_to_plot = NULL,
  colors = NULL
)
}
\arguments{
\item{data}{a dataframe, which was used to estimate the model during hyreg2 or hyreg2_het}

\item{x}{charachter string, column of data to be plotted in x-axis}

\item{y}{charachter string, column of data to be plotted in y-axis}

\item{id_col}{charachter sting, grouping variable, same as was given in model.
if model was estimated without grouping see details}

\item{id_df_model}{dataframe of two columns indicating which group belongs to which class,
first column named as input id_col, second column named "mod_comp".
this variable can easily be filled using the the give_id function, see details.}

\item{type_to_plot}{list of two charachter elements. First: columnname of column containing indicator for type of data,
Second: value of column type, that should be used for the plot, see details of hyreg inputs type and type_cont,type_dich}

\item{colors}{charachter vector, colors to be used in ggplot, default is NULL - than colors are choosen automalically}
}
\value{
ggplot object visulizing x against y by classes from the model
}
\description{
This function can be used to visualize the classification based on the model for different variables.
             ggplot is used.
}
\details{
id_col_df has to be provided anyway, even if the model was estimated without grouping variable.
Since there might be no grouping varibale in the data, we recommend to create a new column "observation"
in data using the rownames/observationnumbers as charachter values and use this column as
input for id_col in plot_hyreg2, additionally you can then use id_df_model =  give_id(data,model,"observation"),
see example below.
}
\examples{
# estimate a model using simulated_data_rnorm

### using | id ####
formula <- y ~  -1 + x1 + x2 + x3 | id
k <- 2
stv <- setNames(c(0.2,0.2,0.2,1,1),c(colnames(simulated_data_norm)[3:5],c("sigma","theta")))
control <- list(iter.max = 1000, verbose = 4)

hyflex_mod <- hyreg2(formula = formula,
                    data =  simulated_data_norm,
                    type =  simulated_data_norm$type,
                    stv = stv,
                    k = k,
                    type_cont = "TTO",
                    type_dich = "DCE_A",
                    opt_method = "L-BFGS-B",
                    control = control,
                    latent = "cont",
                    id_col = "id"
)
# plotting the variables id against y
plot_hyreg2(data = simulated_data_norm,
          x = "id",
          y = "y",
          id_col = "id",
          id_df_model = give_id(data = simulated_data_norm,
                                model = hyflex_mod,
                                id = "id"))

# using only TTO data
plot_hyreg2(data = simulated_data_norm,
          x = "id",
          y = "y",
          id_col = "id",
          id_df_model = give_id(data = simulated_data_norm,
                                model = hyflex_mod,
                                id = "id"),
           type_to_plot = list("type","TTO"))



### model without control for id during | id ###
formula <- y ~  -1 + x1 + x2 + x3
k <- 2
stv <- setNames(c(0.2,0.2,0.2,1,1),c(colnames(simulated_data_norm)[3:5],c("sigma","theta")))
control <- list(iter.max = 1000, verbose = 4)

hyflex_mod <- hyreg2(formula = formula,
                    data =  simulated_data_norm,
                    type =  simulated_data_norm$type,
                    stv = stv,
                    k = k,
                    type_cont = "TTO",
                    type_dich = "DCE_A",
                    opt_method = "L-BFGS-B",
                    control = control,
                    latent = "both"
)

# plotting the variables x2 against y

simulated_data_norm_plot <- simulated_data_norm
simulated_data_norm_plot$observation <- rownames(simulated_data_norm_plot)
plot_hyreg2(data = simulated_data_norm_plot,
          x = "x2",
          y = "y",
          id_col = "observation",
          id_df_model = give_id(data = simulated_data_norm_plot,
                                model = hyflex_mod,
                                id_col = "observation"))


}
\author{
Svenja Elkenkamp & John Grosser
}
