% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyreg2.R
\name{summary_hyreg2}
\alias{summary_hyreg2}
\title{summary_hyreg2}
\usage{
summary_hyreg2(object)
}
\arguments{
\item{object}{modelobject generated with hyreg2 or hyreg2_het}
}
\value{
summary object of bbmle::mle2 model
}
\description{
get model parameters of model generated by hyreg2 oder hyreg2_het
}
\examples{
formula <- y ~  -1 + x1 + x2 + x3 | id
k <- 2
stv <- setNames(c(0.2,0,1,1,1),c(colnames(simulated_data_norm)[3:5],c("sigma","theta")))
control = list(iter.max = 1000, verbose = 4)
rm(counter)
mod <- hyreg2(formula = formula,
                    data =  simulated_data_norm,
                    type =  simulated_data_norm$type,
                    stv = stv,
                    k = k,
                    type_cont = "TTO",
                    type_dich = "DCE_A",
                    opt_method = "L-BFGS-B",
                    control = control,
                    latent = "both",
                    id_col = "id"
)
summary_hyreg2(mod)

}
\author{
Kim Rand & Svenja Elkenkamp
}
