% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_envirnoment.R
\docType{data}
\name{the}
\alias{the}
\title{creating environment for package internal objects}
\format{
An object of class \code{environment} of length 1.
}
\usage{
the
}
\description{
creating environment for package internal objects
}
\keyword{datasets}
