% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLXMRhyreg.R
\name{FLXMRhyreg}
\alias{FLXMRhyreg}
\title{FLXMRhyreg: M-step driver to be used in flexmix}
\usage{
FLXMRhyreg(
  formula = . ~ .,
  family = c("hyreg"),
  type = NULL,
  type_cont = NULL,
  type_dich = NULL,
  variables_both = NULL,
  variables_cont = NULL,
  variables_dich = NULL,
  stv = NULL,
  offset = NULL,
  opt_method = "BFGS",
  optimizer = "optim",
  lower = -Inf,
  upper = Inf,
  ...
)
}
\arguments{
\item{formula}{linear model formula}

\item{family}{default = "hyreg", needed for flexmix::flexmix}

\item{type}{vector containing the indicator whether that datapoint (row)
contains continuous or dichotomous data, see details of hyreg2::hyreg2}

\item{type_cont}{value of type referring to continuous data, see details of hyreg2::hyreg2}

\item{type_dich}{value of type referring to dichotomous data,  see details of hyreg2::hyreg2}

\item{variables_both}{character vector; variables to be fitted on both continuous and dichotomous data.
If not specified, all variables from formula are used. If provided and not all variables from formula
are included, variables_cont and variables_dich must be provided as well, while one of them can be NULL,
 see details of hyreg2::hyreg2}

\item{variables_cont}{character vector; variables to be fitted only on continous data. If provided,
variables_both and variables_dich must be provided as well.}

\item{variables_dich}{character vactor; variables to be fitted only on dichotomous data, if provided,
variables_both and variables_cont must be provided as well.}

\item{stv}{named vector or list of named vectors containing start values for all coefficients
formula, including sigma and theta, see details of hyreg2::hyreg2}

\item{offset}{offset as in flexmix::flexmix}

\item{opt_method}{optimization method to be used in optimizer, default = "BFGS"}

\item{optimizer}{optimizer to be used in bbmle::mle2, default = "optim"}

\item{lower}{default = -INF, lower bound for censored data. If this is used, opt_method must be
set to "L-BFGS-B"}

\item{upper}{default = INF, upper bound for censored data. If this is used, opt_method must be
set to "L-BFGS-B",}

\item{...}{additional arguments for flexmix::flexmix or bbmle::mle2}
}
\value{
a model object, that can be used in hyreg2 as input for parameter model in flexmix::flexmix
}
\description{
Function used in flexmix M-Step to estimate hybrid model
}
\examples{

formula <- y ~  -1 + x1 + x2 + x3
k <- 2
stv <- setNames(c(0.2,0,1,1,1),c(colnames(simulated_data_norm)[3:5],c("sigma","theta")))
rm(counter)



x <- model.matrix(formula,simulated_data_norm)
y <- simulated_data_norm$y
w <- 1
model <- FLXMRhyreg(formula = formula,
                    family=c("hyreg"),
                    type =  simulated_data_norm$type,
                    stv = stv,
                    k = k,
                    type_cont = "TTO",
                    type_dich = "DCE_A",
                    opt_method = "L-BFGS-B",
                    control = list(iter.max = 1000, verbose = 4),
                    offset = NULL,
                    optimizer = "optim",
                    variables_both =  names(stv)[!is.element(names(stv),c("sigma","theta"))],
                    variables_cont = NULL,
                    variables_dich = NULL,
                    lower = -Inf,
                    upper = Inf,
)
}
\author{
Svenja Elkenkamp & Kim Rand
}
