% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLXMRhyreg_het.R
\name{FLXMRhyreg_het}
\alias{FLXMRhyreg_het}
\title{FLXMRhyreg}
\usage{
FLXMRhyreg_het(
  formula = . ~ .,
  family = c("hyreg"),
  type = NULL,
  type_cont = NULL,
  type_dich = NULL,
  variables_both = NULL,
  variables_cont = NULL,
  variables_dich = NULL,
  stv = NULL,
  stv_sigma = NULL,
  offset = NULL,
  opt_method = "BFGS",
  optimizer = "optim",
  lower = -Inf,
  upper = Inf,
  ...
)
}
\arguments{
\item{formula}{Model formula}

\item{family}{default = "hyreg"}

\item{type}{a vector containing an indicator wheter that datapoint contains to TTO or DCE Data}

\item{type_cont}{indicator for continous data}

\item{type_dich}{indicator for dichotoums data}

\item{variables_both}{character vactor; variables to be fitted on TTO and DCE data, if not specified all variables from formula are used}

\item{variables_cont}{character vactor; variables to be fitted only on TTO data}

\item{variables_dich}{character vactor; variables to be fitted only on DCE data}

\item{stv}{named vector or list of named vactors containing start values,
has to be a vector if the same start values should be used for all latent classes,
has to be a list of named vectores if different start values are assumed for the latent classes
has to include start values for sigma and theta as well}

\item{stv_sigma}{vector containing start values for sigma estimates}

\item{offset}{offset as in flexmix}

\item{opt_method}{optimization method to be used in optimizer, default = "BFGS"}

\item{optimizer}{optimizer to be used in bbmle::mle2, default = "optim"}

\item{lower}{opt_method must be set to "L-BFGS-B", lower bound for censored data}

\item{upper}{opt_method must be set to "L-BFGS-B", upper bound for censored data}

\item{...}{additional arguments for flexmix or  bbmle::mle2}
}
\value{
a model with hybrid likelihood
}
\description{
Function used in flexmix M-Step to estimate hybrid model accounting for heteroscedastisity
}
\examples{
Put Example here
}
\author{
Kim Rand & Svenja Elkenkamp
}
