% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLXMRhyreg_het.R
\name{FLXMRhyreg_het}
\alias{FLXMRhyreg_het}
\title{FLXMRhyreg_het: M-step driver to be used in flexmix accounting for heteroscedastisity}
\usage{
FLXMRhyreg_het(
  formula = . ~ .,
  formula_sigma = formula_sigma,
  data,
  family = c("hyreg"),
  type = NULL,
  type_cont = NULL,
  type_dich = NULL,
  variables_both = NULL,
  variables_cont = NULL,
  variables_dich = NULL,
  stv = NULL,
  stv_sigma = NULL,
  offset = NULL,
  opt_method = "BFGS",
  optimizer = "optim",
  lower = -Inf,
  upper = Inf,
  ...
)
}
\arguments{
\item{formula}{Model formula}

\item{formula_sigma}{formula for estimation of sigma to account for heteroscedasticity}

\item{data}{dataset to be used}

\item{family}{default = "hyreg"}

\item{type}{a vector containing an indicator wheter that datapoint contains to TTO or DCE Data}

\item{type_cont}{indicator for continous data}

\item{type_dich}{indicator for dichotoums data}

\item{variables_both}{character vactor; variables to be fitted on TTO and DCE data, if not specified all variables from formula are used}

\item{variables_cont}{character vactor; variables to be fitted only on TTO data}

\item{variables_dich}{character vactor; variables to be fitted only on DCE data}

\item{stv}{named vector or list of named vactors containing start values,
has to be a vector if the same start values should be used for all latent classes,
has to be a list of named vectores if different start values are assumed for the latent classes
has to include start values for sigma and theta as well}

\item{stv_sigma}{vector containing start values for sigma estimates}

\item{offset}{offset as in flexmix}

\item{opt_method}{optimization method to be used in optimizer, default = "BFGS"}

\item{optimizer}{optimizer to be used in bbmle::mle2, default = "optim"}

\item{lower}{opt_method must be set to "L-BFGS-B", lower bound for censored data}

\item{upper}{opt_method must be set to "L-BFGS-B", upper bound for censored data}

\item{...}{additional arguments for flexmix or  bbmle::mle2}
}
\value{
a model with hybrid likelihood
}
\description{
Function used in flexmix M-Step to estimate hybrid model accounting for heteroscedastisity
}
\examples{

formula <- y ~  -1 + x1 + x2 + x3
formula_sigma <- y ~ x1 + x2 + x3
stv <- setNames(c(0.2,0,1,1),c(colnames(simulated_data_norm)[3:5],c("theta")))
stv_sigma <- setNames(c(0.2,0.2,0.1,1),c(colnames(simulated_data_norm)[3:5],c("(Intercept)")))


x <- model.matrix(formula,simulated_data_norm)
y <- simulated_data_norm$y
w <- 1
model <- FLXMRhyreg_het(formula = formula,
                     formula_sigma = formula_sigma,
                    family=c("hyreg"),
                    type =  simulated_data_norm$type,
                    stv = stv,
                    stv_sigma = stv_sigma,
                    k = k,
                    type_cont = "TTO",
                    type_dich = "DCE_A",
                    opt_method = "L-BFGS-B",
                    control = list(iter.max = 1000, verbose = 4),
                    offset = NULL,
                    optimizer = "optim",
                    variables_both =  names(stv)[!is.element(names(stv),c("theta"))],
                    variables_cont = NULL,
                    variables_dich = NULL,
                    lower = -Inf,
                    upper = Inf,
)
}
\author{
Kim Rand & Svenja Elkenkamp
}
