% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyreg2.R
\name{hyreg2}
\alias{hyreg2}
\title{hyreg2}
\usage{
hyreg2(
  formula,
  data,
  type,
  type_cont,
  type_dich,
  k = 1,
  control = NULL,
  stv = NULL,
  offset = NULL,
  opt_method = "BFGS",
  optimizer = "optim",
  lower = -Inf,
  upper = Inf,
  latent = "both",
  id_col = NULL,
  variables_both = NULL,
  variables_dich = NULL,
  variables_cont = NULL,
  ...
)
}
\arguments{
\item{formula}{Model formula}

\item{data}{a dataframe containing the data}

\item{type}{a vector containing an indicator wheter that datapoint contains to TTO or DCE Data}

\item{type_cont}{indicator for continous data}

\item{type_dich}{indicator for dichotoums data}

\item{k}{numeric, number of latent classes to be estimated via flexmix::flexmix}

\item{control}{control vector for flexmix::flexmix}

\item{stv}{named vector or list of named vactors containing start values,
has to be a vector if the same start values should be used for all latent classes,
has to be a list of named vectores if different start values are assumed for the latent classes
has to include start values for sigma and theta as well}

\item{offset}{offset as in flexmix}

\item{opt_method}{optimization method to be used in optimizer, default = "BFGS"}

\item{optimizer}{optimizer to be used in bbmle::mle2, default = "optim"}

\item{lower}{opt_method must be set to "L-BFGS-B", lower bound for censored data}

\item{upper}{opt_method must be set to "L-BFGS-B", upper bound for censored data}

\item{latent}{one of "both","cont" or "dich", see details}

\item{id_col}{character, column name containing participant ids}

\item{variables_both}{character vactor; variables to be fitted on TTO and DCE data, if not specified all variables from formula are used}

\item{variables_dich}{character vactor; variables to be fitted only on DCE data}

\item{variables_cont}{character vactor; variables to be fitted only on TTO data}

\item{...}{additional arguments for flexmix::flexmix or bbmle::mle2}
}
\value{
model of type flemix
}
\description{
Estimation of hybrid model for EQ-5D data
}
\examples{
Put Example here
}
\author{
Kim Rand & Svenja Elkenkamp
}
