% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyreg2_het.R
\name{hyreg2_het}
\alias{hyreg2_het}
\title{hyreg2_het: function for model estimation for EQ5D valueset data accounting for heteroscedasticity in TTO}
\usage{
hyreg2_het(
  formula,
  formula_sigma = NULL,
  data,
  type,
  type_cont,
  type_dich,
  k = 1,
  control = NULL,
  stv = NULL,
  stv_sigma = NULL,
  offset = NULL,
  opt_method = "BFGS",
  optimizer = "optim",
  lower = -Inf,
  upper = Inf,
  latent = "both",
  id_col = NULL,
  classes_only = FALSE,
  variables_both = NULL,
  variables_dich = NULL,
  variables_cont = NULL,
  ...
)
}
\arguments{
\item{formula}{linear model formula}

\item{formula_sigma}{formula for sigma estimation, if not provided formula is taken without |}

\item{data}{a dataframe containing the data}

\item{type}{a vector containing an indicator wheter that datapoint contains to TTO or DCE Data}

\item{type_cont}{indicator for continous data, part of type}

\item{type_dich}{indicator for dichotoums data,  part of type}

\item{k}{numeric, number of latent classes to be estimated via flexmix::flexmix}

\item{control}{control vector for flexmix::flexmix}

\item{stv}{named vector or list of named vactors containing start values for all coefficients from formula,
has to be a vector if the same start values should be used for all latent classes,
has to be a list of named vectores if different start values are assumed for the latent classes
has to include start values for sigma and theta as well and it is important to use the names "theta" and "sigma"
Using colnames(model.matrix(formula,data)) (formula without |) you can check, which variables need a stv value.}

\item{stv_sigma}{named vector with start values for sigma. Have to be the same variables as given in formula_sigma. Should NOT end with _h}

\item{offset}{offset as in flexmix}

\item{opt_method}{optimization method to be used in optimizer, default = "BFGS"}

\item{optimizer}{optimizer to be used in bbmle::mle2, default = "optim"}

\item{lower}{opt_method must be set to "L-BFGS-B", lower bound for censored data}

\item{upper}{opt_method must be set to "L-BFGS-B", upper bound for censored data}

\item{latent}{one of "both","cont" or "dich", see details}

\item{id_col}{character, column name containing participant ids, needed if latent != "both}

\item{classes_only}{logical, default FALSE, indicate if only classification should be done without estimating model parameters
only possible for latent = "cont" or "dich"}

\item{variables_both}{character vactor; variables to be fitted on TTO and DCE data, if not specified all variables from formula are used}

\item{variables_dich}{character vactor; variables to be fitted only on DCE data, if provided variables_both and variables_cont have to be provided as well.
Then all variables from stv have to be part in exactly one of the vectors variables_cont, variables_dich and variables_both.
A vector can also be set to NULL while the others contain all variables}

\item{variables_cont}{character vactor; variables to be fitted only on TTO data, if provided variables_both and variables_dich have to be provided as well.
Then all variables from stv have to be part in exactly one of the vectors variables_cont, variables_dich and variables_both.
A vector can also be set to NULL while the others contain all variables}

\item{...}{additional arguments for flexmix::flexmix or bbmle::mle2}
}
\value{
model of type flemix, coefficients named ..._h are coefficients for heteroscedasticity
}
\description{
Estimation of hybrid model for EQ-5D data
}
\details{
Using input latent = "both" means, that latent classes are estimated on both parts of the data: TTO and DCE.
Using latent = "cont" the latent classes are estimated only on the TTO part and afterwards with these classes the combined likelihood of normal and binomial distribution
is used to estimate the model parameters. Using latent = "dich" the latent classes are estimated only on the DCE partand afterwards with these classes the combined likelihood of normal and binomial distribution
is used to estimate the model parameters. id_col has to be specified and only people with observations in both parts of the data are taken into account, hence it can happen that some
ids are removed from the estimation.


For datasets individuals with more than one observation it can be useful to estimate the latent classes for the individuals instead of for each datapoint;
use | idcolumn in the formula.
}
\examples{
formula <- y ~  -1 + x1 + x2 + x3
formula_sigma <- y ~  x1 + x2 + x3

k <- 1
stv <- setNames(c(0.2,0,1,1),c(colnames(simulated_data_norm)[3:5],c("theta")))
stv_sigma <- setNames(c(0.2,0,1,1),c(colnames(simulated_data_norm)[3:5],c("(Intercept)")))
control = list(iter.max = 1000, verbose = 4)
rm(counter)
mod <- hyreg2_het(formula = formula,
                   formula_sigma = formula_sigma,
                    data =  simulated_data_norm,
                    type =  simulated_data_norm$type,
                    stv = stv,
                    stv_sigma = stv_sigma,
                    k = k,
                    type_cont = "TTO",
                    type_dich = "DCE_A",
                    opt_method = "L-BFGS-B",
                    control = control,
                    latent = "both",
                    id_col = "id"
)
summary_hyreg2(mod)
}
\author{
Svenja Elkenkamp & Kim Rand
}
