% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{simulated_data}
\format{
\subsection{\code{simulated_data}}{

A simulated data frame with 480 rows and 25 columns, following a combination of normal and binomial distribution
\describe{
\item{type}{type of data}
\item{y}{result of the model y = -1 + mo2 + mo3 + ... + ad4 + ad5}
\item{mo2, mo3, mo4, mo5,
sc2, sc3, sc4, sc5,
ua2, ua3, ua4, ua5,
pd2, pd3, pd4, pd5,,
ad2, ad3, ad4, ad5,}{dummy variables for EQ5D data simulation}
\item{class}{original class of the data point}
\item{id}{id number of observations to simulated different persons}
\item{y_cens}{column y censored at 2 (upper boundary)}
...
}
}
}
\source{
{simulated with true parameter values:
Class 1: sigma  = XXX, theta = XXX and c(mo2,mo3,mo4,mo5, XXX) = c(XXX)
Class 2: sigma  = XXX, theta = XXX and c(mo2,mo3,mo4,mo5, XXX) = c(XXX)}
}
\usage{
simulated_data
}
\description{
simulated_data
}
\keyword{datasets}
