% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyreg2.R
\name{summary_hyreg2}
\alias{summary_hyreg2}
\title{summary_hyreg2}
\usage{
summary_hyreg2(object)
}
\arguments{
\item{object}{modelobject generated with hyreg2}
}
\value{
summary object of bbmle::mle2 model
}
\description{
get model parameters of hyreg2 model
}
\examples{
Put Example here
}
\author{
Kim Rand & Svenja Elkenkamp
}
